Added a Sleep() call to the Windows network extensions code for the case where there are no open descriptors.  This was causing 100% cpu utilization. Fixed a spurious cast reported by Shentino.

+++ src/parser/grammar.c
@@ -247,8 +247,8 @@ static int gramtok(string *str, ssizet *strlen, char *buffer, unsigned int *bufl
 		    }
 		    node[thisnode].type = RGX_PAREN;
 
-		    nstack[paren++] = (short) topnode;
-		    nstack[paren++] = (short) thisnode;
+		    nstack[paren++] = topnode;
+		    nstack[paren++] = thisnode;
 		    topnode = thisnode = -1;
 		    break;
